Updated avro schema
Finished first draft of the url_filter script
It queries the url topic for messages that are already in the sytem fo filter away those if it should try to reproduce them.
It auto dismisses the urls that are no longer relevants and commits to advance the offset in the consumer group for tha url topic
It does not commit a valid url until all data is successfully filtered from the input and produced to the output topic with forecast data.

added commit() function to consumer in kafka_clients.py
and flush() to producer

added lookup dictionary for filter_url in utils.py

created job scheduler functions to centralise adding of the same jobs for all instances where the same job should be added.